types files distribution improved
+ added building.types.ts
+ added emitter.types.ts
+ added idb.types.ts
+ added modules.types.ts
+ added savegame.types.ts
+ added ttcst.types.ts
+ added utils.types.ts
/ improved types.ts
/ improved mission..types.ts
/ improved type dependent files
- removed unused code/files/dir
