+++ src/app/page.tsx
@@ -1,8 +1,7 @@
 'use client'
 
 import { useEffect } from "react";
-import Landing from "./shared/components/landing/landing";
-import Sidebar from "./shared/components/nav/sidenav";
+import Landing from "@shared//components/landing";
 import { init } from "./script/setup";
 
 export const API_KEY = process.env.TOMTOM_API_KEY
@@ -16,7 +15,6 @@ export default function page() {
 
     return (
         <>
-            <Sidebar />
             <Landing />
         </>
     )
