+++ src/app/script/setup.ts
@@ -1,10 +1,16 @@
+import { map_inst } from "@shared/components/map";
+
 import { invoke } from "@tauri-apps/api/tauri";
-import { GameEmitter } from "../emitter";
-import { BuildingInterface, DatabaseGetOptions, MissionInterface, PresenceInterface } from "../shared/types/types";
+
+import tt, { LngLatBounds } from "@tomtom-international/web-sdk-maps";
+
+import { GameEmitter } from "./emitter";
 import { getDB } from "../indexed_db";
-import { map_inst } from "../shared/components/map";
-import tt, { LngLatBounds, Marker } from "@tomtom-international/web-sdk-maps";
 import { Mission, generateMissionData } from "./gen/mission";
+
+import { BuildingInterface } from "@shared/types/building.types";
+import { DatabaseGetOptions } from "@shared/types/idb.types";
+import { MissionInterface } from "@shared/types/missions.types";
 
 // import { currentMonitor, appWindow, PhysicalPosition } from "@tauri-apps/api/window";
 
@@ -48,7 +54,7 @@ export function init() {
     //         throw new Error(e);
     //     })
 
-    GameEmitter.on('EVENT_GAME_START', async (data) => {
+    GameEmitter.on('EVENT_GAME_START', async (data: any) => {
 
         const getFromDBOptions: DatabaseGetOptions = {
             database: 'DB_SAVEGAME_DATA',
@@ -132,7 +138,7 @@ export function init() {
 
                 marker.setLngLat(mission.location.coords);
                 marker.addTo(map_inst);
-                popup.setHTML(`<strong>${mission.mission.specific}</strong><br>${mission.location.free_address}`);
+                popup.setHTML(`<strong>${mission.mission}</strong><br>${mission.location.free_address}`);
                 popup.addTo(map_inst);
                 marker.setPopup(popup);
                 marker.togglePopup();
@@ -167,7 +173,7 @@ export function init() {
 
                         marker.setLngLat(newMission.data.location.coords);
                         marker.addTo(map_inst);
-                        popup.setHTML(`<strong>${newMission.data.mission.specific}</strong><br>${newMission.data.location.free_address}`);
+                        popup.setHTML(`<strong>${newMission.data.mission}</strong><br>${newMission.data.location.free_address}`);
                         popup.addTo(map_inst);
                         marker.setPopup(popup);
                         marker.togglePopup();
