--- src/app/shared/assets/names.json
@@ -1,211 +1,0 @@
-{
-    "first_names": [
-        "Lena",
-        "Niklas",
-        "Emilia",
-        "Felix",
-        "Hannah",
-        "David",
-        "Sophia",
-        "Maximilian",
-        "Johanna",
-        "Timo",
-        "Laura",
-        "Sebastian",
-        "Marie",
-        "Tobias",
-        "Tobi",
-        "Torben",
-        "Lisa",
-        "Benjamin",
-        "Anna",
-        "Simon",
-        "Lina",
-        "Lukas",
-        "Julia~",
-        "Jan",
-        "Lea",
-        "Philipp",
-        "Katharina",
-        "Christian",
-        "Clara",
-        "Cindy~",
-        "Jonas",
-        "Nora",
-        "Alexander",
-        "Mia",
-        "Luca",
-        "Elias",
-        "Sarah",
-        "Tim",
-        "Vanessa",
-        "Julian",
-        "Caroline",
-        "Daniel",
-        "Franziska",
-        "Paulina",
-        "Paula",
-        "Paul",
-        "Emma",
-        "Olivia",
-        "Ava",
-        "Sophia",
-        "Isabella",
-        "Mia",
-        "Charlotte",
-        "Amelia",
-        "Harper",
-        "Evelyn",
-        "Yuna",
-        "Sakura",
-        "Mei",
-        "Airi",
-        "Sora",
-        "Santiago",
-        "Mateo",
-        "Sebastian",
-        "Benjamin",
-        "Noah",
-        "Lucas",
-        "Ethan",
-        "Alexander",
-        "William",
-        "Michael",
-        "James",
-        "Daniel",
-        "David",
-        "Joseph",
-        "Christopher",
-        "Anthony",
-        "Matthew",
-        "Andrew",
-        "Joshua",
-        "Jacob",
-        "Ryan",
-        "Brandon",
-        "Tyler",
-        "Hannah",
-        "Emily",
-        "Chloe",
-        "Aria",
-        "Abigail",
-        "Elizabeth",
-        "Grace",
-        "Ella",
-        "Victoria",
-        "Natalie",
-        "Isabella",
-        "Sophia"
-    ],
-    "last_names": [
-        "Müller",
-        "Schmidt",
-        "Schneider",
-        "Fischer",
-        "Weber",
-        "Meyer",
-        "Wagner",
-        "Becker",
-        "Schulz",
-        "Hoffmann",
-        "Schäfer",
-        "Koch",
-        "Bauer",
-        "Richter",
-        "Klein",
-        "Wolf",
-        "Neumann",
-        "Schwarz",
-        "Zimmermann",
-        "Braun",
-        "Lange",
-        "Krause",
-        "Werner",
-        "Schmidt",
-        "Schneider",
-        "Schultz",
-        "Schmitt",
-        "Klein",
-        "Meier",
-        "Fischer",
-        "Wagner",
-        "Schuster",
-        "Vogel",
-        "Stein",
-        "Jung",
-        "Roth",
-        "Böhmer",
-        "Köhler",
-        "Maier",
-        "Lehmann",
-        "Schmid",
-        "Berger",
-        "Novak",
-        "Kovacs",
-        "Szabo",
-        "Horvath",
-        "Balogh",
-        "Varga",
-        "Takacs",
-        "Papp",
-        "Farkas",
-        "Toth",
-        "Nagy",
-        "Kovalev",
-        "Ivanov",
-        "Kuznetsov",
-        "Sokolov",
-        "Popov",
-        "Orlov",
-        "Kozlov",
-        "Mikhailov",
-        "Smith",
-        "Johnson",
-        "Garcia",
-        "Gonzalez",
-        "Kim",
-        "Lee",
-        "Park",
-        "Wang",
-        "Chen",
-        "Wu",
-        "Yamamoto",
-        "Tanaka",
-        "Suzuki",
-        "Takahashi",
-        "Ito",
-        "Hernandez",
-        "Lopez",
-        "Perez",
-        "Rodriguez",
-        "Gomez",
-        "Martinez",
-        "Sanchez",
-        "Ramirez",
-        "Jackson",
-        "Williams",
-        "Jones",
-        "Brown",
-        "Davis",
-        "Wilson",
-        "Taylor",
-        "Anderson",
-        "Johnson",
-        "Clark",
-        "Thomas",
-        "Moore",
-        "Allen",
-        "Baker",
-        "Carter",
-        "Mitchell",
-        "Nguyen",
-        "Robinson",
-        "Scott",
-        "Turner",
-        "Walker",
-        "Young",
-        "Zhang",
-        "Li",
-        "Liu"
-    ]
-}
