--- src/app/shared/components/nav/sidenav.css
@@ -1,33 +1,0 @@
-.sidebar {
-    position: absolute;
-    top: 0;
-    bottom: 0;
-    left: 0;
-    width: 4rem;
-    background-color: var(--primary);
-    z-index: 2;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: space-between;
-}
-
-button.sidebar-item {
-    position: relative;
-    all: unset;
-    height: 4rem;
-    width: 4rem;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    cursor: pointer;
-    transition: background-color .25s ease-in-out, border .25s ease-in-out;
-}
-
-button.sidebar-item.back-btn {
-    border-bottom-right-radius: 1rem;
-}
-
-.sidebar-item:hover {
-    background-color: var(--primary-600);
-}
