+++ src/app/shared/style/loading.css
