+++ src/app/shared/style/titlebar.css
